Fix to wait a short time right after the start
There was a issue that if you hit enter key to start the program in the
terminal, a new line would be typed all the way through until you hit
the key again.
This was because all input from the keyboard was grabbed before the
enter key was released, and therefore the event that the key was
released was discarded.

Therefore, a small wait time was inserted between the start of the
program and the grabbing the keyboard.