Fix scrolling speed for easier adjustment
The scrolling speed was reasonably fast even when set to the minimum
value of 1, so it was difficult to adjust it to be slower.

Therefore, REL_WHEEL and REL_HWHEEL were changed to the *_HI_RES
version so that the set value would be larger for the same speed
compared to the older version.
This now made it easier to set a slower scrolling speed.