Change config file format to .xml file
Until now, the config file has been actually just a header file.
In consideration of allowing more flexible configuration or adding
new features, the format of the config file was switched to .xml file.

A config file template can be generated by config/gen_config. To change
settings, edit the template. The way to give the config file to
marionette is to specify it as a command line argument. marionette
validates its config file in the initial stage, using
config/config.dtd.

libxml2 is used to manipulate the .xml files.